# docs/swagger.yaml
definitions:
  domain.AnalysisStatus:
    enum:
    - processing
    - completed
    - failed
    type: string
    x-enum-varnames:
    - Processing
    - Completed
    - Failed
  domain.Gender:
    enum:
    - male
    - female
    type: string
    x-enum-varnames:
    - GenderMale
    - GenderFemale
  domain.Ingredient:
    properties:
      description:
        type: string
      id:
        type: string
      name:
        type: string
    type: object
  domain.Recommendation:
    properties:
      description:
        type: string
      id:
        type: string
      title:
        type: string
    type: object
  domain.SkinType:
    enum:
    - oily
    - dry
    - normal
    - combination
    type: string
    x-enum-varnames:
    - Oily
    - Dry
    - Normal
    - Combination
  dto.AnalysisDetailResponse:
    properties:
      analysis_data:
        additionalProperties: {}
        type: object
      created_at:
        type: string
      file_url:
        type: string
      id:
        type: string
      ingredients:
        items:
          $ref: '#/definitions/domain.Ingredient'
        type: array
      photo_id:
        type: string
      recommendations:
        items:
          $ref: '#/definitions/domain.Recommendation'
        type: array
      skin_type:
        $ref: '#/definitions/domain.SkinType'
      status:
        $ref: '#/definitions/domain.AnalysisStatus'
      updated_at:
        type: string
    type: object
  dto.AnalysisListItemResponse:
    properties:
      created_at:
        type: string
      file_url:
        type: string
      id:
        type: string
      photo_id:
        type: string
      skin_type:
        $ref: '#/definitions/domain.SkinType'
      status:
        $ref: '#/definitions/domain.AnalysisStatus'
      updated_at:
        type: string
    type: object
  dto.AuthResponse:
    properties:
      access_token:
        type: string
      refresh_token:
        type: string
      token_type:
        type: string
    type: object
  dto.CreateAnalysisResponse:
    properties:
      file_url:
        type: string
      id:
        type: string
      photo_id:
        type: string
      status:
        $ref: '#/definitions/domain.AnalysisStatus'
      uploaded_at:
        type: string
    type: object
  dto.ErrorResponse:
    properties:
      code:
        type: string
      message:
        type: string
    type: object
  dto.HealthResponse:
    properties:
      status:
        type: string
    type: object
  dto.LoginRequest:
    properties:
      email:
        example: user@example.com
        type: string
      password:
        type: string
    required:
    - email
    - password
    type: object
  dto.LogoutRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.MessageResponse:
    properties:
      message:
        type: string
    type: object
  dto.RefreshRequest:
    properties:
      refresh_token:
        type: string
    required:
    - refresh_token
    type: object
  dto.RegisterRequest:
    properties:
      birth_date:
        example: "2000-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/domain.Gender'
        enum:
        - male
        - female
      name:
        type: string
      password:
        type: string
    required:
    - email
    - name
    - password
    type: object
  dto.UserPasswordRequest:
    properties:
      current_password:
        type: string
      new_password:
        type: string
    required:
    - current_password
    - new_password
    type: object
  dto.UserRequest:
    properties:
      birth_date:
        example: "2000-01-01T00:00:00Z"
        type: string
      email:
        example: user@example.com
        type: string
      gender:
        allOf:
        - $ref: '#/definitions/domain.Gender'
        enum:
        - male
        - female
      name:
        type: string
    type: object
  dto.UserResponse:
    properties:
      birth_date:
        type: string
      created_at:
        type: string
      email:
        type: string
      gender:
        $ref: '#/definitions/domain.Gender'
      id:
        type: string
      name:
        type: string
      updated_at:
        type: string
    type: object
info:
  contact: {}
  title: Backend API
  version: "1.0"
paths:
  /analyses:
    get:
      description: Returns paginated list of user analyses
      parameters:
      - default: 10
        description: Limit
        in: query
        name: limit
        type: integer
      - default: 0
        description: Offset
        in: query
        name: offset
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/dto.AnalysisListItemResponse'
            type: array
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get user analysis history
      tags:
      - analyses
    post:
      consumes:
      - multipart/form-data
      description: Upload photo and start skin analysis
      parameters:
      - description: Photo to analyze
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            $ref: '#/definitions/dto.CreateAnalysisResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Create analysis from photo
      tags:
      - analyses
  /analyses/{id}:
    delete:
      description: Deletes user analysis by id
      parameters:
      - description: Analysis ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete analysis by ID
      tags:
      - analyses
    get:
      description: Returns detailed analysis result
      parameters:
      - description: Analysis ID
        in: path
        name: id
        required: true
        type: string
      - default: 10
        description: Recommendations limit
        in: query
        name: recs_limit
        type: integer
      - default: 10
        description: Ingredients limit
        in: query
        name: ings_limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AnalysisDetailResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "403":
          description: Forbidden
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get analysis by ID
      tags:
      - analyses
  /auth/login:
    post:
      consumes:
      - application/json
      description: Authenticates user and returns access and refresh tokens
      parameters:
      - description: Login request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LoginRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Login user
      tags:
      - auth
  /auth/logout:
    post:
      consumes:
      - application/json
      description: Invalidates refresh token and logs user out from current session
      parameters:
      - description: Logout request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.LogoutRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.MessageResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Logout user
      tags:
      - auth
  /auth/refresh:
    post:
      consumes:
      - application/json
      description: Issues new access and refresh tokens using refresh token
      parameters:
      - description: Refresh request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RefreshRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Refresh tokens
      tags:
      - auth
  /auth/register:
    post:
      consumes:
      - application/json
      description: Creates new user and returns JWT tokens
      parameters:
      - description: Register request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.RegisterRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/dto.AuthResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      summary: Register user
      tags:
      - auth
  /health:
    get:
      description: Returns OK if service is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
      summary: Check service health
      tags:
      - health
  /health/db:
    get:
      description: Returns OK if db is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.HealthResponse'
      summary: Check db health
      tags:
      - health
  /health/ml:
    get:
      description: Returns status of ML service availability
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.HealthResponse'
      summary: Check ML service health
      tags:
      - health
  /health/storage:
    get:
      description: Returns OK if storage is running
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.HealthResponse'
        "503":
          description: Service Unavailable
          schema:
            $ref: '#/definitions/dto.HealthResponse'
      summary: Check storage health
      tags:
      - health
  /users/me:
    delete:
      consumes:
      - application/json
      description: Deletes authenticated user account
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Delete current user
      tags:
      - users
    get:
      consumes:
      - application/json
      description: Returns information about the authenticated user
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Get current user
      tags:
      - users
    patch:
      consumes:
      - application/json
      description: Updates current user and returns updated user information
      parameters:
      - description: User Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/dto.UserResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update current user
      tags:
      - users
  /users/me/password:
    patch:
      consumes:
      - application/json
      description: Changes current user password after verifying current password
      parameters:
      - description: User Password Request
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/dto.UserPasswordRequest'
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "401":
          description: Unauthorized
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/dto.ErrorResponse'
      security:
      - BearerAuth: []
      summary: Update password
      tags:
      - users
securityDefinitions:
  BearerAuth:
    description: 'Enter token as: Bearer <your_token>'
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"